requirements.txt
jieba
scikit-learn
matplotlib
numpy
pandas
openpyxl
google-generativeai
rouge-score
time